添加了基于内容推荐

+++ cajr-provider-news/src/main/java/com/cajr/util/MapValueComparator.java
@@ -7,9 +7,9 @@ import java.util.Map;
  * @Author CAJR
  * @create 2020/2/18 22:36
  */
-public class MapValueComparator implements Comparator<Map.Entry<Long,Double>> {
+public class MapValueComparator implements Comparator<Map.Entry<Integer,Double>> {
     @Override
-    public int compare(Map.Entry<Long, Double> o1, Map.Entry<Long, Double> o2) {
+    public int compare(Map.Entry<Integer, Double> o1, Map.Entry<Integer, Double> o2) {
         return o1.getValue().compareTo(o2.getValue());
     }
 }
